Split rewrite operators over different files
